[PVR] Refactor CPVRGUIActions step 2: Move recording related actions to own class.
