+++ xbmc/pvr/guilib/PVRGUIActions.h
@@ -9,6 +9,7 @@
 #pragma once
 
 #include "pvr/PVRChannelNumberInputHandler.h"
+#include "pvr/guilib/PVRGUIActionsRecordings.h"
 #include "pvr/guilib/PVRGUIActionsTimers.h"
 #include "pvr/guilib/PVRGUIChannelNavigator.h"
 #include "pvr/settings/PVRSettings.h"
@@ -40,7 +41,6 @@ namespace PVR
 
   class CPVRChannel;
   class CPVRChannelGroupMember;
-  class CPVRRecording;
   class CPVRStreamProperties;
   class CPVRTimerInfoTag;
 
@@ -65,7 +65,7 @@ namespace PVR
     void SwitchToPreviousChannel();
   };
 
-  class CPVRGUIActions : public CPVRGUIActionsTimers
+  class CPVRGUIActions : public CPVRGUIActionsRecordings, public CPVRGUIActionsTimers
   {
   public:
     CPVRGUIActions();
@@ -91,54 +91,6 @@ namespace PVR
      * @return true on success, false otherwise.
      */
     bool FindSimilar(const std::shared_ptr<CFileItem>& item) const;
-
-    /*!
-     * @brief Open a dialog with information for a given recording.
-     * @param item containing a recording.
-     * @return true on success, false otherwise.
-     */
-    bool ShowRecordingInfo(const std::shared_ptr<CFileItem>& item) const;
-
-    /*!
-     * @brief Open the recording settings dialog to edit a recording.
-     * @param item containing the recording to edit.
-     * @return true on success, false otherwise.
-     */
-    bool EditRecording(const std::shared_ptr<CFileItem>& item) const;
-
-    /*!
-     * @brief Check if any recording settings can be edited.
-     * @param item containing the recording to edit.
-     * @return true on success, false otherwise.
-     */
-    bool CanEditRecording(const CFileItem& item) const;
-
-    /*!
-     * @brief Delete a recording, always showing a confirmation dialog.
-     * @param item containing a recording to delete.
-     * @return true, if the recording was deleted successfully, false otherwise.
-     */
-    bool DeleteRecording(const std::shared_ptr<CFileItem>& item) const;
-
-    /*!
-     * @brief Delete all watched recordings contained in the given folder, always showing a confirmation dialog.
-     * @param item containing a recording folder containing the items to delete.
-     * @return true, if the recordings were deleted successfully, false otherwise.
-     */
-    bool DeleteWatchedRecordings(const std::shared_ptr<CFileItem>& item) const;
-
-    /*!
-     * @brief Delete all recordings from trash, always showing a confirmation dialog.
-     * @return true, if the recordings were permanently deleted successfully, false otherwise.
-     */
-    bool DeleteAllRecordingsFromTrash() const;
-
-    /*!
-     * @brief Undelete a recording.
-     * @param item containing a recording to undelete.
-     * @return true, if the recording was undeleted successfully, false otherwise.
-     */
-    bool UndeleteRecording(const std::shared_ptr<CFileItem>& item) const;
 
     /*!
      * @brief Get a localized resume play label, if the given item can be resumed.
@@ -367,33 +319,6 @@ namespace PVR
   private:
     CPVRGUIActions(const CPVRGUIActions&) = delete;
     CPVRGUIActions const& operator=(CPVRGUIActions const&) = delete;
-
-    /*!
-     * @brief Open a dialog to confirm to delete a recording.
-     * @param item the recording to delete.
-     * @return true, to proceed with delete, false otherwise.
-     */
-    bool ConfirmDeleteRecording(const std::shared_ptr<CFileItem>& item) const;
-
-    /*!
-     * @brief Open a dialog to confirm delete all watched recordings contained in the given folder.
-     * @param item containing a recording folder containing the items to delete.
-     * @return true, to proceed with delete, false otherwise.
-     */
-    bool ConfirmDeleteWatchedRecordings(const std::shared_ptr<CFileItem>& item) const;
-
-    /*!
-     * @brief Open a dialog to confirm to permanently remove all deleted recordings.
-     * @return true, to proceed with delete, false otherwise.
-     */
-    bool ConfirmDeleteAllRecordingsFromTrash() const;
-
-    /*!
-     * @brief Open the recording settings dialog.
-     * @param recording containing the recording the settings shall be displayed for.
-     * @return true, if the dialog was ended successfully, false otherwise.
-     */
-    bool ShowRecordingSettings(const std::shared_ptr<CPVRRecording>& recording) const;
 
     /*!
      * @brief Check whether resume play is possible for a given item, display "resume from ..."/"play from start" context menu in case.
